The great elementscontext merge!
