+++ packages/core/src/index.ts
@@ -10,6 +10,7 @@ export * from './constants';
 export * from './contenttype';
 export * from './crypto';
 export * from './default-values';
+export * from './elements-context';
 export * from './eventtarget';
 export * from './fhircast';
 export * from './fhirlexer/parse';
